requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.7.1
psycopg2-binary==2.9.10
bcrypt==4.2.1
python-dotenv==1.0.0
flask-cors==5.0.1
boto3==1.34.75
Pillow==11.1.0
python-magic==0.4.27
numpy==2.2.3
opencv-python==4.11.0.86
celery==5.4.0
redis==5.2.1
Flask-Migrate==4.0.5